Small entity refactor in preparation for user
